requirements.txt
streamlit==1.28.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
firebase-admin==6.2.0
python-dotenv==1.0.0
google-generativeai==0.2.2